cleanup purity macros disable plotting for now removed Performance/printPurity.C (all purity calculations will be done in one macro)

+++ Histogramming/PhotonPurity.h
@@ -18,7 +18,7 @@ class PhotonPurity {
     double sigshift;
     double sigshifterr;
     double chisq;
-    double ndf;
+    int ndf;
 
     double purity;
